[0.3] Remote storage support storage in SDK and CLI (#302)
## Notable changes

- see `CHANGELOG.md`

## References

- resolves https://github.com/Talus-Network/nexus/issues/516
- resolves https://github.com/Talus-Network/nexus/issues/518
- resolves https://github.com/Talus-Network/nexus/issues/517
- resolves https://github.com/Talus-Network/nexus/issues/520
- resolves https://github.com/Talus-Network/nexus/issues/519

## Checklist

- [x] keep a changelog
- [x] write tests
- [ ] ~~create tickets for `TODO: <>` statements~~
- [x] create or update documentation

+++ tools/storage-walrus/src/upload_file.rs
@@ -48,13 +48,13 @@ pub(crate) struct Input {
     publisher_url: Option<String>,
     /// The number of epochs to store the file
     #[serde(default = "default_epochs")]
-    epochs: u64,
+    epochs: u8,
     /// Optional address to which the created Blob object should be sent
     #[serde(default)]
     send_to: Option<String>,
 }
 
-fn default_epochs() -> u64 {
+fn default_epochs() -> u8 {
     1
 }
