+++ tools/storage-walrus/src/upload_json.rs
@@ -43,13 +43,13 @@ pub(crate) struct Input {
     aggregator_url: Option<String>,
     /// Number of epochs to store the data
     #[serde(default = "default_epochs")]
-    epochs: u64,
+    epochs: u8,
     /// Optional address to which the created Blob object should be sent
     #[serde(default)]
     send_to_address: Option<String>,
 }
 
-fn default_epochs() -> u64 {
+fn default_epochs() -> u8 {
     1
 }
